[tex] Simplify math mode snippets.
When I wanted to add more snippets for latex math mode commands there
was a possibility that they had conflicts with the previously added
ones. Also, I had to add a new snippet for each command. With this
change, now every time I type two alphanumeric characters after a space,
a pipe, a double pipe or an opening parenthesis or square bracket, it
adds a "\" character to start a new latex command.

Furthermore, to expand a complex expression like an summation, now I
just have to press tab after the corresponding latex command to activate
the snippet.